registerCount=1..4 と gain の小数対応

+++ app/src/main/java/com/example/meterdemo/meter/model/MeterPoint.kt
@@ -4,21 +4,21 @@ data class MeterPoint(
     val name: String,
     val address: Int,
     val registerCount: Int,
-    val gain: Int,
+    val gain: Double,
     val dataType: DataType,
     val unit: String = "",
     val initialRawValue: Int,
     val wordByteOrder: WordByteOrder = WordByteOrder.MSB_MSB
 ) {
     fun formattedValue(rawValue: Int): String {
-        return if (gain <= 1) {
+        return if (gain <= 1.0) {
             if (unit.isBlank()) {
                 rawValue.toString()
             } else {
                 "$rawValue $unit"
             }
         } else {
-            val scaled = rawValue.toDouble() / gain.toDouble()
+            val scaled = rawValue.toDouble() / gain
             if (unit.isBlank()) {
                 scaled.toString()
             } else {
